+++ zpages/templates.go
@@ -44,16 +44,16 @@ type headerData struct {
 func parseTemplate(name string) *template.Template {
 	f, err := internal.FS(false).Open("/templates/" + name + ".html")
 	if err != nil {
-		log.Panicf("%v: %v", name, err)
+		log.Panicf("%v: %v", name, err) // nolint: revive  // Called during initialization.
 	}
 	defer func() {
 		if err = f.Close(); err != nil {
-			log.Panicf("%v: %v", name, err)
+			log.Panicf("%v: %v", name, err) // nolint: revive  // Called during initialization.
 		}
 	}()
 	text, err := ioutil.ReadAll(f)
 	if err != nil {
-		log.Panicf("%v: %v", name, err)
+		log.Panicf("%v: %v", name, err) // nolint: revive  // Called during initialization.
 	}
 	return template.Must(template.New(name).Funcs(templateFunctions).Parse(string(text)))
 }